Added Path Following Behavior.
Modified the Missile so that it could follow a path using the waypoints
given to it.  It also skips over waypoints if they are within the
minimum seek distance of the current position.  The behavior is pretty
good when the missile is lined up on the path to start.  If not, it can
"circle" the points, which brings it close to the next point while
still circling, so it circles the entire path until it ends.  Overall,
though, the behavior is pretty good for a missile and the general seek
behavior will work well if the waypoints are spread out.

The main scene has been modified to demonstrate all behaviors.